+++ Controllers/MediaController.m
@@ -195,7 +195,7 @@ static MediaController *sharedController = nil;
 
 - (id) copyWithZone:(NSZone *)zone								{ return self; }
 - (id) retain													{ return self; }
-- (unsigned) retainCount										{ return UINT_MAX;  /* denotes an object that cannot be released */ }
+- (NSUInteger) retainCount										{ return UINT_MAX;  /* denotes an object that cannot be released */ }
 - (void) release												{ /* do nothing */ }
 - (id) autorelease												{ return self; }
